refactor: Sections / 기존 관리자 객체를 삭제하고 Sections로 기능 통합
