+++ src/main/java/subway/application/SectionService.java
@@ -9,8 +9,6 @@ import subway.dao.StationDao;
 import subway.domain.Line;
 import subway.domain.PathFinder;
 import subway.domain.Section;
-import subway.domain.SectionAddManager;
-import subway.domain.SectionRemoveManager;
 import subway.domain.Sections;
 import subway.domain.ShortestPath;
 import subway.domain.Station;
@@ -39,11 +37,11 @@ public class SectionService {
         Sections sections = sectionDao.findAllBy(line);
         int distance = request.getDistance();
 
-        SectionAddManager sectionAddManager = new SectionAddManager(sections);
-        sectionAddManager.validate(upStation, downStation, distance);
+        sections.validateAddition(upStation, downStation, distance);
         Section section = new Section(line, upStation, downStation, distance);
-        sectionAddManager.lookForChange(section)
-            .ifPresent(sectionDao::update);
+        if (sections.isAddedInMiddle(section)) {
+            sectionDao.update(sections.findSectionToChange(section));
+        }
         sectionDao.save(section);
     }
 
@@ -53,10 +51,10 @@ public class SectionService {
         Sections sections = sectionDao.findAllBy(line);
         Station station = getStationBy(stationId);
 
-        SectionRemoveManager sectionRemoveManager = new SectionRemoveManager(sections);
-        sectionRemoveManager.validate(station);
-        sectionRemoveManager.lookForChange(station)
-            .ifPresent(sectionDao::update);
+        sections.validateRemoval(station);
+        if (sections.isInMiddle(station)) {
+            sectionDao.update(sections.findSectionToChange(station));
+        }
         sectionDao.deleteByLineAndStation(line, station);
     }
