+++ src/main/java/subway/domain/Section.java
@@ -27,7 +27,7 @@ public class Section {
 
     private void validatePositive(int distance) {
         if (distance <= 0) {
-            throw new IllegalArgumentException("구간 길이는 양수여야합니다 distance: \"" + distance + "\"");
+            throw new IllegalArgumentException("구간 길이는 양수여야 합니다 distance: \"" + distance + "\"");
         }
     }
