+++ src/main/java/subway/domain/Sections.java
@@ -13,6 +13,8 @@ import org.springframework.util.CollectionUtils;
 
 public class Sections {
 
+    private static final int MIN_SECTION_COUNT = 1;
+
     private final List<Section> sections;
 
     public Sections(List<Section> candidateValues) {
@@ -83,34 +85,108 @@ public class Sections {
             .collect(Collectors.toSet());
     }
 
-    public List<Station> getStations() {
-        List<Station> stations = sections.stream()
-            .map(Section::getUpStation)
-            .collect(Collectors.toList());
-        stations.add(getLast().getDownStation());
-        return stations;
+    public void validateAddition(final Station upStation, final Station downStation, final int distance) {
+        validateLineHasOneOf(upStation, downStation);
+
+        filter(matchOneOf(upStation, downStation))
+            .ifPresent(section -> validateDistance(section, distance));
     }
 
-    public boolean hasStation(Station station) {
+    private void validateDistance(final Section section, final int distance) {
+        if (section.isNotLongerThan(distance)) {
+            throw new IllegalArgumentException("추가할 구간의 크기가 너무 큽니다.");
+        }
+    }
+
+    public boolean isAddedInMiddle(Section newSection) {
+        return filter(section -> section.matchEitherStation(newSection)).isPresent();
+    }
+
+    public Section findSectionToChange(Section newSection) {
+        return filter(section -> section.matchEitherStation(newSection))
+            .map(section -> section.cutBy(newSection))
+            .orElseThrow(() -> new IllegalArgumentException("구간 추가로 인해 변경된 구간이 존재하지 않습니다."));
+    }
+
+    private Predicate<Section> matchOneOf(final Station upStation, final Station downStation) {
+        return section -> section.hasUpStation(upStation) || section.hasDownStation(downStation);
+    }
+
+    private void validateLineHasOneOf(final Station upStation, final Station downStation) {
+        boolean hasUpStation = hasStation(upStation);
+        boolean hasDownStation = hasStation(downStation);
+
+        validateBoth(hasUpStation, hasDownStation);
+        validateNotBoth(hasUpStation, hasDownStation);
+    }
+
+    private void validateBoth(final boolean hasUpStation, final boolean hasDownStation) {
+        if (hasUpStation && hasDownStation) {
+            throw new IllegalArgumentException("두 역이 모두 노선에 포함되어 있습니다.");
+        }
+    }
+
+    private void validateNotBoth(final boolean hasUpStation, final boolean hasDownStation) {
+        if (!hasUpStation && !hasDownStation) {
+            throw new IllegalArgumentException("두 역이 모두 노선에 포함되어 있지 않습니다.");
+        }
+    }
+
+    public void validateRemoval(final Station station) {
+        validateSize();
+        validateStationInLine(station);
+    }
+
+    private void validateSize() {
+        if (sections.size() <= MIN_SECTION_COUNT) {
+            throw new IllegalStateException("노선의 구간이 " + MIN_SECTION_COUNT + "개인 경우 삭제할 수 없습니다.");
+        }
+    }
+
+    private void validateStationInLine(final Station station) {
+        if (!hasStation(station)) {
+            throw new IllegalArgumentException("노선에 등록되어 있지 않은 역은 제거할 수 없습니다.");
+        }
+    }
+
+    public boolean isInMiddle(final Station station) {
+        return !isAtEndOfLine(station);
+    }
+
+    public Section findSectionToChange(final Station station) {
+        if (isAtEndOfLine(station)) {
+            throw new IllegalArgumentException("구간 삭제로 인해 변경된 구간이 존재하지 않습니다.");
+        }
+
+        Section upperSection = getUpperSection(station);
+        Section lowerSection = getLowerSection(station);
+        return upperSection.extendBy(lowerSection);
+    }
+
+    private boolean isAtEndOfLine(final Station station) {
+        return isFirst(station) || isLast(station);
+    }
+
+    private Section getUpperSection(final Station station) {
+        return filter(section -> section.hasDownStation(station))
+            .orElseThrow(() -> new IllegalStateException("역의 위 구간을 찾을 수 없습니다."));
+    }
+
+    private Section getLowerSection(final Station station) {
+        return filter(section -> section.hasUpStation(station))
+            .orElseThrow(() -> new IllegalStateException("역의 아래 구간을 찾을 수 없습니다."));
+    }
+
+    private boolean hasStation(Station station) {
         return getStations().contains(station);
     }
 
-    public boolean isFirst(Station station) {
+    private boolean isFirst(Station station) {
         return getFirst().hasUpStation(station);
     }
 
-    public boolean isLast(Station station) {
+    private boolean isLast(Station station) {
         return getLast().hasDownStation(station);
-    }
-
-    public Optional<Section> filter(Predicate<Section> condition) {
-        return sections.stream()
-            .filter(condition)
-            .findAny();
-    }
-
-    public int getSize() {
-        return sections.size();
     }
 
     private Section getFirst() {
@@ -120,6 +196,20 @@ public class Sections {
 
     private Section getLast() {
         return CollectionUtils.lastElement(sections);
+    }
+
+    private Optional<Section> filter(Predicate<Section> condition) {
+        return sections.stream()
+            .filter(condition)
+            .findAny();
+    }
+
+    public List<Station> getStations() {
+        List<Station> stations = sections.stream()
+            .map(Section::getUpStation)
+            .collect(Collectors.toList());
+        stations.add(getLast().getDownStation());
+        return stations;
     }
 
     public List<Section> getSections() {
